+++ src/stase/__init__.py
@@ -38,4 +38,4 @@ __all__ = [
     "fieldSignificance_FDR",
 ]
 
-__version__ = "0.6.2"
+__version__ = "0.6.3"